# geochronos-backend/verification/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoChronos Application API",
    "description": "Production-quality local FastAPI backend integrating GeoChronos Data Foundation and GeoChronos Intelligence Engine for air-gapped Earth Observation analytics (SIH26227).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/features": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "List Features",
        "operationId": "list_features_api_features_get",
        "parameters": [
          {
            "name": "feature_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Feature class filter",
              "title": "Feature Class"
            },
            "description": "Feature class filter"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Year filter (2011, 2015, 2020, 2026)",
              "title": "Year"
            },
            "description": "Year filter (2011, 2015, 2020, 2026)"
          },
          {
            "name": "observation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Observation ID filter",
              "title": "Observation Id"
            },
            "description": "Observation ID filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "include_geojson",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to include GeoJSON FeatureCollection",
              "default": false,
              "title": "Include Geojson"
            },
            "description": "Whether to include GeoJSON FeatureCollection"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/temporal/analyze": {
      "post": {
        "tags": [
          "Temporal Analysis"
        ],
        "summary": "Analyze Temporal",
        "operationId": "analyze_temporal_api_temporal_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemporalAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemporalAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/temporal/compare": {
      "post": {
        "tags": [
          "Temporal Analysis"
        ],
        "summary": "Compare Temporal",
        "operationId": "compare_temporal_api_temporal_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemporalCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Compare Temporal Api Temporal Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/map/features": {
      "get": {
        "tags": [
          "Map"
        ],
        "summary": "Get Map Features",
        "operationId": "get_map_features_api_map_features_get",
        "parameters": [
          {
            "name": "feature_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Feature class filter",
              "title": "Feature Class"
            },
            "description": "Feature class filter"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Year (2011, 2015, 2020, 2026)",
              "default": 2026,
              "title": "Year"
            },
            "description": "Year (2011, 2015, 2020, 2026)"
          },
          {
            "name": "min_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box min longitude (WGS84)",
              "title": "Min Lon"
            },
            "description": "Bounding box min longitude (WGS84)"
          },
          {
            "name": "min_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box min latitude (WGS84)",
              "title": "Min Lat"
            },
            "description": "Bounding box min latitude (WGS84)"
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box max longitude (WGS84)",
              "title": "Max Lon"
            },
            "description": "Bounding box max longitude (WGS84)"
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box max latitude (WGS84)",
              "title": "Max Lat"
            },
            "description": "Bounding box max latitude (WGS84)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/map/feature/{feature_id}": {
      "get": {
        "tags": [
          "Map"
        ],
        "summary": "Get Single Feature Geojson",
        "operationId": "get_single_feature_geojson_api_map_feature__feature_id__get",
        "parameters": [
          {
            "name": "feature_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feature Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeature"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeatureItem": {
        "properties": {
          "feature_id": {
            "type": "string",
            "title": "Feature Id"
          },
          "feature_class": {
            "type": "string",
            "title": "Feature Class"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "observation_id": {
            "type": "string",
            "title": "Observation Id"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source_item_id": {
            "type": "string",
            "title": "Source Item Id"
          },
          "cube_time_index": {
            "type": "integer",
            "title": "Cube Time Index"
          },
          "area_m2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "feature_id",
          "feature_class",
          "year",
          "observation_id",
          "confidence",
          "source_item_id",
          "cube_time_index"
        ],
        "title": "FeatureItem"
      },
      "FeatureListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureItem"
            },
            "type": "array",
            "title": "Features"
          },
          "geojson": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "features"
        ],
        "title": "FeatureListResponse"
      },
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJSONGeometry"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "geometry"
        ],
        "title": "GeoJSONFeature"
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          },
          "total_features": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Features"
          },
          "crs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crs"
          }
        },
        "type": "object",
        "title": "GeoJSONFeatureCollection"
      },
      "GeoJSONGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "GeoJSONGeometry"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemporalAnalysisRequest": {
        "properties": {
          "feature_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Class",
            "description": "Class of feature, e.g. water_body, vegetation, urban_area"
          },
          "from_year": {
            "type": "integer",
            "title": "From Year",
            "description": "Base comparison year",
            "default": 2011
          },
          "to_year": {
            "type": "integer",
            "title": "To Year",
            "description": "Target comparison year",
            "default": 2026
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Optional spatial area or locality"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent",
            "description": "Change intent (e.g. feature_change, feature_disappearance, feature_appearance)",
            "default": "feature_change"
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          }
        },
        "type": "object",
        "title": "TemporalAnalysisRequest"
      },
      "TemporalAnalysisResponse": {
        "properties": {
          "from_year": {
            "type": "integer",
            "title": "From Year"
          },
          "to_year": {
            "type": "integer",
            "title": "To Year"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "summary_statistics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary Statistics"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "results_count": {
            "type": "integer",
            "title": "Results Count"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "geojson": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              },
              {
                "type": "null"
              }
            ]
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "from_year",
          "to_year",
          "intent",
          "summary_statistics",
          "limitations",
          "results_count",
          "results",
          "provenance"
        ],
        "title": "TemporalAnalysisResponse"
      },
      "TemporalCompareRequest": {
        "properties": {
          "from_year": {
            "type": "integer",
            "title": "From Year",
            "default": 2011
          },
          "to_year": {
            "type": "integer",
            "title": "To Year",
            "default": 2026
          },
          "feature_classes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Classes"
          }
        },
        "type": "object",
        "title": "TemporalCompareRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}